+++ src/Repositories/SqlStatementRepository.php
@@ -43,26 +43,26 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
         SqlStatementKeys::CATEGORIES =>
             'SELECT t0.*,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_varchar t2
+                       FROM ${table:eav_attribute} t1, ${table:catalog_category_entity_varchar} t2
                       WHERE t1.attribute_code = \'name\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS name,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_varchar t2
+                       FROM ${table:eav_attribute t1}, ${table:catalog_category_entity_varchar} t2
                       WHERE t1.attribute_code = \'url_path\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS url_path
-               FROM catalog_category_entity AS t0',
+               FROM ${table:catalog_category_entity} AS t0',
         SqlStatementKeys::CATEGORY_VARCHARS_BY_ENTITY_IDS =>
             'SELECT t1.*
-               FROM catalog_category_entity AS t0
-         INNER JOIN catalog_category_entity_varchar AS t1
+               FROM ${table:catalog_category_entity} AS t0
+         INNER JOIN ${table:catalog_category_entity_varchar} AS t1
                  ON t1.row_id = t0.row_id
-         INNER JOIN eav_attribute AS t2
+         INNER JOIN ${table:eav_attribute} AS t2
                  ON t2.entity_type_id = 3
                 AND t2.attribute_code = \'name\'
                 AND t1.attribute_id = t2.attribute_id
@@ -70,87 +70,87 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                 AND t0.entity_id IN (?)',
         SqlStatementKeys::CATEGORY_VARCHARS_BY_ROW_IDS =>
             'SELECT t1.*
-               FROM catalog_category_entity AS t0
-         INNER JOIN catalog_category_entity_varchar AS t1
+               FROM ${table:catalog_category_entity} AS t0
+         INNER JOIN ${table:catalog_category_entity_varchar} AS t1
                  ON t1.row_id = t0.row_id
-         INNER JOIN eav_attribute AS t2
+         INNER JOIN ${table:eav_attribute} AS t2
                  ON t2.entity_type_id = 3
                 AND t2.attribute_code = \'name\'
                 AND t1.attribute_id = t2.attribute_id
                 AND t1.store_id = 0
                 AND t0.row_id IN (?)',
         SqlStatementKeys::CATEGORY =>
-            'SELECT * FROM catalog_category_entity WHERE row_id = :row_id',
+            'SELECT * FROM ${table:catalog_category_entity} WHERE row_id = :row_id',
         SqlStatementKeys::CATEGORY_DATETIMES =>
             'SELECT *
-               FROM catalog_category_entity_datetime
+               FROM ${table:catalog_category_entity_datetime}
               WHERE row_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_DECIMALS =>
             'SELECT *
-               FROM catalog_category_entity_decimal
+               FROM ${table:catalog_category_entity_decimal}
               WHERE row_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_INTS =>
             'SELECT *
-               FROM catalog_category_entity_int
+               FROM ${table:catalog_category_entity_int}
               WHERE row_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_TEXTS =>
             'SELECT *
-               FROM catalog_category_entity_text
+               FROM ${table:catalog_category_entity_text}
               WHERE row_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_VARCHARS =>
             'SELECT *
-               FROM catalog_category_entity_varchar
+               FROM ${table:catalog_category_entity_varchar}
               WHERE row_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::CATEGORY_DATETIMES_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_datetime t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_datetime} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.row_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_DECIMALS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_decimal t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_decimal} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.row_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_INTS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_int t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_int} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.row_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_TEXTS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_text t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_text} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.row_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CATEGORY_VARCHARS_BY_PK_AND_STORE_ID =>
             'SELECT t0.*,
                     t1.attribute_code
-               FROM catalog_category_entity_varchar t0
-         INNER JOIN eav_attribute t1
+               FROM ${table:catalog_category_entity_varchar} t0
+         INNER JOIN ${table:eav_attribute} t1
                  ON t1.attribute_id = t0.attribute_id
               WHERE t0.row_id = :pk
                 AND t0.store_id = :store_id',
         SqlStatementKeys::CREATE_SEQUENCE_CATEGORY =>
-            'INSERT INTO sequence_catalog_category VALUES ()',
+            'INSERT INTO ${table:sequence_catalog_category} VALUES ()',
         SqlStatementKeys::CREATE_CATEGORY =>
             'INSERT
-               INTO catalog_category_entity
+               INTO ${table:catalog_category_entity}
                     (attribute_set_id,
                      entity_id,
                      created_in,
@@ -174,7 +174,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                      :level,
                      :children_count)',
         SqlStatementKeys::UPDATE_CATEGORY =>
-            'UPDATE catalog_category_entity
+            'UPDATE ${table:catalog_category_entity}
                 SET attribute_set_id = :attribute_set_id,
                     entity_id = :entity_id,
                     created_in = :created_in,
@@ -189,7 +189,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
               WHERE entity_id = :entity_id',
         SqlStatementKeys::CREATE_CATEGORY_DATETIME =>
             'INSERT
-               INTO catalog_category_entity_datetime
+               INTO ${table:catalog_category_entity_datetime}
                     (row_id,
                      attribute_id,
                      store_id,
@@ -199,7 +199,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                     :store_id,
                     :value)',
         SqlStatementKeys::UPDATE_CATEGORY_DATETIME =>
-            'UPDATE catalog_category_entity_datetime
+            'UPDATE ${table:catalog_category_entity_datetime}
                SET row_id = :row_id,
                    attribute_id = :attribute_id,
                    store_id = :store_id,
@@ -207,7 +207,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
              WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_DECIMAL =>
             'INSERT
-               INTO catalog_category_entity_decimal
+               INTO ${table:catalog_category_entity_decimal}
                     (row_id,
                      attribute_id,
                      store_id,
@@ -217,7 +217,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                     :store_id,
                     :value)',
         SqlStatementKeys::UPDATE_CATEGORY_DECIMAL =>
-            'UPDATE catalog_category_entity_decimal
+            'UPDATE ${table:catalog_category_entity_decimal}
                 SET row_id = :row_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -225,7 +225,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_INT =>
             'INSERT
-               INTO catalog_category_entity_int
+               INTO ${table:catalog_category_entity_int}
                     (row_id,
                      attribute_id,
                      store_id,
@@ -235,7 +235,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_INT =>
-            'UPDATE catalog_category_entity_int
+            'UPDATE ${table:catalog_category_entity_int}
                 SET row_id = :row_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -243,7 +243,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_VARCHAR =>
             'INSERT
-               INTO catalog_category_entity_varchar
+               INTO ${table:catalog_category_entity_varchar}
                     (row_id,
                      attribute_id,
                      store_id,
@@ -253,7 +253,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_VARCHAR =>
-            'UPDATE catalog_category_entity_varchar
+            'UPDATE ${table:catalog_category_entity_varchar}
                 SET row_id = :row_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -261,7 +261,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_CATEGORY_TEXT =>
             'INSERT
-               INTO catalog_category_entity_text
+               INTO ${table:catalog_category_entity_text}
                     (row_id,
                      attribute_id,
                      store_id,
@@ -271,7 +271,7 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_CATEGORY_TEXT =>
-            'UPDATE catalog_category_entity_text
+            'UPDATE ${table:catalog_category_entity_text}
                 SET row_id = :row_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -280,17 +280,17 @@ class SqlStatementRepository extends \TechDivision\Import\Category\Repositories\
     );
 
     /**
-     * Initialize the the SQL statements.
+     * Initializes the SQL statement repository with the primary key and table prefix utility.
+     *
+     * @param \IteratorAggregate<\TechDivision\Import\Utils\SqlCompilerInterface> $compilers The array with the compiler instances
      */
-    public function __construct()
+    public function __construct(\IteratorAggregate $compilers)
     {
 
-        // call the parent constructor
-        parent::__construct();
+        // pass primary key + table prefix utility to parent instance
+        parent::__construct($compilers);
 
-        // merge the class statements
-        foreach ($this->statements as $key => $statement) {
-            $this->preparedStatements[$key] = $statement;
-        }
+        // compile the SQL statements
+        $this->compile($this->statements);
     }
 }